refactor(microservices): move tcp socket logic to an abstract class
